Completed Footer Social

+++ app/Http/Controllers/Admin/BrandController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\DataTables\Admin\BrandDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Brand;
+use App\Models\Product;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -109,10 +110,15 @@ class BrandController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id){
+    public function destroy(string $id)
+    {
         $brand = Brand::findOrFail($id);
+        if (Product::where('brand_id', $brand->id)->count() > 0) {
+            return response(['status' => 'error', 'message' => 'This brand have products you can\'t delete it.']);
+        }
+        $this->deleteImage($brand->logo);
         $brand->delete();
 
-        return response(['status'=>'success','message'=>'Deleted Successfully']);
+        return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
 }
